refactor(epicenter): implement platform-specific exports with conditional imports
This refactors the epicenter package to use Node.js conditional exports (subpath exports) for proper platform-specific behavior. The key change is that `EpicenterConfig.storageDir` is now only present in Node environments and completely absent from the browser type.

The implementation follows a consistent pattern across all modules:
- `.shared.ts`: Platform-agnostic logic and base types
- `.browser.ts`: Browser-specific implementation (no filesystem access)
- `.node.ts`: Node-specific implementation (with filesystem/storageDir support)

Package exports are configured so that bundlers automatically resolve the correct variant based on the target platform, with Node as the default fallback.
