+++ packages/epicenter/package.json
@@ -1,14 +1,24 @@
 {
 	"name": "@epicenter/hq",
 	"version": "0.0.1",
-	"main": "./src/index.ts",
-	"types": "./src/index.ts",
+	"main": "./src/index.node.ts",
+	"types": "./src/index.node.ts",
 	"exports": {
-		".": "./src/index.ts",
+		".": {
+			"browser": "./src/index.browser.ts",
+			"node": "./src/index.node.ts",
+			"default": "./src/index.node.ts"
+		},
+		"./providers/hocuspocus": "./src/providers/hocuspocus.ts",
+		"./providers/persistence": {
+			"browser": "./src/providers/persistence/index.browser.ts",
+			"node": "./src/providers/persistence/index.node.ts",
+			"default": "./src/providers/persistence/index.node.ts"
+		},
 		"./cli": "./src/cli/index.ts",
-		"./providers": "./src/providers/index.ts",
 		"./indexes/markdown": "./src/indexes/markdown/index.ts",
-		"./indexes/sqlite": "./src/indexes/sqlite/index.ts"
+		"./indexes/sqlite": "./src/indexes/sqlite/index.ts",
+		"./server": "./src/server/index.ts"
 	},
 	"bin": {
 		"epicenter": "./src/cli/bin.ts"
