+++ packages/epicenter/src/core/epicenter/index.ts
@@ -1,7 +1,8 @@
-// Definition (config side)
+/**
+ * Default epicenter exports (Node.js).
+ *
+ * Internal code (CLI, server, tests) imports from this file.
+ * These are all Node-only contexts, so we re-export the node version.
+ */
 
-export type { ActionInfo, EpicenterClient } from './client';
-// Runtime (client side)
-export { createEpicenterClient, iterActions } from './client';
-export type { EpicenterConfig } from './config';
-export { defineEpicenter } from './config';
+export * from './index.node';
