+++ packages/epicenter/src/core/workspace/index.node.ts
@@ -1,0 +1,20 @@
+/**
+ * Node.js-specific workspace exports.
+ *
+ * This file is selected by bundlers when the "node" condition is matched.
+ */
+
+// Provider types (shared)
+export type { Provider, ProviderContext } from '../provider';
+// Runtime - node version
+export { createWorkspaceClient } from './client.node';
+// Types from shared (no platform-specific code)
+export type { WorkspaceClient, WorkspacesToClients } from './client.shared';
+
+// Config types and definition (shared)
+export type {
+	AnyWorkspaceConfig,
+	WorkspaceConfig,
+	WorkspacesToExports,
+} from './config';
+export { defineWorkspace } from './config';
