+++ packages/epicenter/src/index.shared.ts
@@ -1,0 +1,175 @@
+/**
+ * Epicenter: YJS-First Collaborative Workspace System
+ *
+ * A unified architecture for building self-contained, globally synchronizable workspaces
+ * with real-time collaboration via YJS.
+ *
+ * ## Core Concepts
+ *
+ * - **YJS Document**: Source of truth (CRDT, collaborative)
+ * - **Indexes**: Synchronized snapshots for different query patterns
+ * - **Column Schemas**: Pure JSON definitions (no Drizzle builders)
+ *
+ * ## Data Flow
+ *
+ * Write to YJS → Indexes auto-sync → Query indexes
+ *
+ * This file contains all platform-agnostic exports shared between
+ * browser and Node.js entry points.
+ */
+
+// Re-export commonly used Drizzle utilities for querying indexes
+export {
+	and,
+	asc,
+	desc,
+	eq,
+	gt,
+	gte,
+	inArray,
+	isNotNull,
+	isNull,
+	like,
+	lt,
+	lte,
+	ne,
+	not,
+	or,
+	sql,
+} from 'drizzle-orm';
+
+export type {
+	Action,
+	Mutation,
+	Query,
+	WorkspaceActionMap,
+	WorkspaceExports,
+} from './core/actions';
+
+// Action helpers
+export {
+	defineMutation,
+	defineQuery,
+	defineWorkspaceExports,
+	extractActions,
+	isAction,
+	isMutation,
+	isNamespace,
+	isQuery,
+	walkActions,
+} from './core/actions';
+
+export type { Db, TableHelper } from './core/db/core';
+
+// Database utilities
+export { createEpicenterDb } from './core/db/core';
+
+export type { GetResult, RowResult, YRow } from './core/db/table-helper';
+
+// Epicenter types (shared across platforms)
+// Note: EpicenterConfig and defineEpicenter are exported from platform-specific
+// entry points because they have different type signatures (Node adds storageDir)
+export type {
+	ActionInfo,
+	EpicenterClient,
+} from './core/epicenter/client.shared';
+export { iterActions } from './core/epicenter/client.shared';
+
+export type {
+	EpicenterOperationError,
+	IndexError,
+	ValidationError,
+} from './core/errors';
+
+// Error types
+export {
+	EpicenterOperationErr,
+	IndexErr,
+	ValidationErr,
+} from './core/errors';
+
+export type {
+	Index,
+	IndexContext,
+	IndexExports,
+	WorkspaceIndexMap,
+} from './core/indexes';
+
+// Index system
+export { defineIndexExports } from './core/indexes';
+
+export type { Provider, ProviderContext } from './core/provider';
+
+export type {
+	BooleanColumnSchema,
+	CellValue,
+	ColumnSchema,
+	ColumnType,
+	DateColumnSchema,
+	DateIsoString,
+	DateWithTimezoneString,
+	Id,
+	IdColumnSchema,
+	IntegerColumnSchema,
+	JsonColumnSchema,
+	PartialSerializedRow,
+	RealColumnSchema,
+	Row,
+	SelectColumnSchema,
+	SerializedCellValue,
+	SerializedRow,
+	TableSchema,
+	TableValidators,
+	TagsColumnSchema,
+	TextColumnSchema,
+	TimezoneId,
+	WorkspaceSchema,
+	WorkspaceValidators,
+	YtextColumnSchema,
+} from './core/schema';
+
+// Column schema system
+export {
+	boolean,
+	createTableValidators,
+	createWorkspaceValidators,
+	DATE_WITH_TIMEZONE_STRING_REGEX,
+	DateWithTimezone,
+	DateWithTimezoneFromString,
+	date,
+	generateId,
+	ISO_DATETIME_REGEX,
+	id,
+	integer,
+	isDateWithTimezone,
+	isDateWithTimezoneString,
+	isIsoDateTimeString,
+	json,
+	real,
+	select,
+	serializeCellValue,
+	TIMEZONE_ID_REGEX,
+	tags,
+	text,
+	ytext,
+} from './core/schema';
+
+// Core types
+export type { AbsolutePath, EpicenterDir, StorageDir } from './core/types';
+
+// Workspace types (shared across platforms)
+export type {
+	WorkspaceClient,
+	WorkspacesToClients,
+} from './core/workspace/client.shared';
+export type {
+	AnyWorkspaceConfig,
+	WorkspaceConfig,
+	WorkspacesToExports,
+} from './core/workspace/config';
+export { defineWorkspace } from './core/workspace/config';
+
+// Note: Indexes (markdown, sqlite) are NOT re-exported here to avoid bundling
+// Node.js-only code in browser builds. Import them directly from subpaths:
+//   import { markdownIndex } from '@epicenter/hq/indexes/markdown';
+//   import { sqliteIndex } from '@epicenter/hq/indexes/sqlite';
